Fix deadlock due to blocking unsafe FFI call
Using the Distribution metric could cause a deadlock if the following
happened:

 * One thread holds the lock and is in the critical section.

 * Another thread is blocked on the lock (i.e. inside an unsafe FFI
   call).

 * GC strikes.

The GC waits for all unsafe FFI calls to return, but the blocked FFI
never returned and hence the deadlock.

The solution is to do the locking, computation, and unlocking in a
single FFI call. That way all FFI calls will eventually return when GC
happens.

Fixes #24.