cbits: fix min/max combine logic
The internal `combine` function that is used to read the value of a
`Distribution` contains `min` and `max` values, and as the name implies,
combines them in the Semigroup sense: to combine two samples `a` and `b`,
for `max` you simply calculate `max(a->max, b->max)`, likewise for `min`

While this is fine in theory, in practice it is flawed for this case: to
read a metric's current values, you must combine it with an empty
sample, where `min = max = 0`. This logic is incorrect, because any
value that was previously seen that is greater than zero will be
overwritten by zero every time you read the sample. This is a flat out
bug. (Likewise for `max` and values less-than zero).

In more abstract terms, the identity members of the `min/max` monoids
are *not* zero, but `maxBound/minBound`!

However, there's an easier way to go forward here: `combine` is *only*
ever used with a zero structure on the right hand side, and it is
completely internal. Therefore the logic is trivial: just copy the
`max/min` values from the old structure into the new structure, since
there's no point in even doing the comparison anyway.

Signed-off-by: Austin Seipp <aseipp@pobox.com>